Cleaned up code

+++ src/main/java/SevenZip/Compression/LZMA/Base.java
@@ -48,7 +48,7 @@ public class Base {
 	public static final int GetLenToPosState(int len) {
 		len -= kMatchMinLen;
 		if (len < kNumLenToPosStates) return len;
-		return (int) (kNumLenToPosStates - 1);
+		return (kNumLenToPosStates - 1);
 	}
 
 	public static final int	kNumAlignBits										= 4;
