+++ src/main/java/SevenZip/Compression/LZMA/Decoder.java
@@ -260,7 +260,7 @@ public class Decoder {
 		int pb = remainder / 5;
 		int dictionarySize = 0;
 		for (int i = 0; i < 4; i++)
-			dictionarySize += ((int) (properties[1 + i]) & 0xFF) << (i * 8);
+			dictionarySize += ((properties[1 + i]) & 0xFF) << (i * 8);
 		if (!SetLcLpPb(lc, lp, pb)) return false;
 		return SetDictionarySize(dictionarySize);
 	}
